Updated GUI labels
"u12" is the frame number at which you fall on the ground (cross referenced it with Sadamitsu).
"val3" value is always a multiple of 4 except it's default value, which is 1. It loops pushback val3/8 times.
I wanted to try changing the 'value' label in the Extradata Pushback editor too but it seems that would require more than just adding entries in the field labels. Pushback extradata value is the horizontal offset. Centre for that is the player. +ve puts opp. in front of you. -ve behind you.